+++ C02/dummylua-2-4/test/p2_test.c
@@ -13,8 +13,8 @@ void p2_test_main() {
        luaL_pushnil(L);
    }
 
-   int start_time = time(NULL);
-   int end_time = time(NULL);
+   int start_time = (int)time(NULL);
+   int end_time = (int)time(NULL);
    size_t max_bytes = 0;
    struct global_State* g = G(L);
    int j = 0;
@@ -33,7 +33,7 @@ void p2_test_main() {
             printf("timestamp:%d totalbytes:%f kb \n", (int)time(NULL), (float)(g->totalbytes + g->GCdebt) / 1024.0f);
         }
    }
-   end_time = time(NULL);
+   end_time = (int)time(NULL);
    printf("finish test start_time:%d end_time:%d max_bytes:%f kb \n", start_time, end_time, (float)max_bytes / 1024.0f);
 
    luaL_close(L);
